+++ src/functionalTest/groovy/com/autonomousapps/android/AbstractAndroidSpec.groovy
@@ -23,7 +23,7 @@ abstract class AbstractAndroidSpec extends AbstractFunctionalSpec {
   private static final SUPPORTED_AGP_VERSIONS = [
     AgpVersion.version('3.5.3'),
     AgpVersion.version('3.6.3'),
-    AgpVersion.version('4.0.0-rc01'),
+    AgpVersion.version('4.0.0'),
     AgpVersion.version('4.1.0-alpha09')
   ]